test/ScapeGoatTreeTest.kt

import org.junit.jupiter.api.Tag
import org.junit.jupiter.api.Test
import java.util.*
import kotlin.test.assertEquals
import kotlin.test.assertFalse
import kotlin.test.assertTrue

class ScapeGoatTreeTest {
    private fun testAdd(create: () -> CheckableSortedSet<Int>) {
        val tree = create()
        assertEquals(0, tree.size)
        assertFalse(tree.contains(5))
        tree.add(10)
        tree.add(5)
        tree.add(7)
        tree.add(10)
        assertEquals(3, tree.size)
        assertTrue(tree.contains(5))
        tree.add(3)
        tree.add(1)
        tree.add(3)
        tree.add(4)
        assertEquals(6, tree.size)
        assertFalse(tree.contains(8))
        tree.add(8)
        tree.add(15)
        tree.add(15)
        tree.add(20)
        assertEquals(9, tree.size)
        assertTrue(tree.contains(8))
        assertTrue(tree.checkInvariant())
        assertEquals(1, tree.first())
        assertEquals(20, tree.last())

        val random = (167..97561).random()
        tree.add(random)
        assertEquals(10, tree.size)
        assertTrue(tree.contains(random))
        assertEquals(random, tree.last())
    }

    @Test
    @Tag("Example")
    fun testAddSGT() {
        testAdd { createKotlinTree() }
    }

    private fun <T : Comparable<T>> createKotlinTree(): CheckableSortedSet<T> = ScapeGoatTree()

    private fun testRemove(create: () -> CheckableSortedSet<Int>) {
        val random = Random()
        for (iteration in 1..100) {
            val list = mutableListOf<Int>()
            for (i in 1..20) {
                list.add(random.nextInt(100))
            }
            val binarySet = create()
            assertFalse(binarySet.remove(42))
            for (element in list) {
                binarySet += element
            }
            val originalHeight = binarySet.height()
            val toRemove = list[random.nextInt(list.size)]
            val oldSize = binarySet.size

            assertTrue(binarySet.remove(toRemove))
            assertEquals(oldSize - 1, binarySet.size)
            println("Removing $toRemove from $list")
            for (element in list) {
                val inn = element != toRemove
                assertEquals(
                        inn, element in binarySet,
                        "$element should be ${if (inn) "in" else "not in"} tree"
                )
            }
            list.remove(toRemove)

            assertTrue(binarySet.checkInvariant(), "Binary tree invariant is false after tree.remove()")
            assertTrue(
                    binarySet.height() <= originalHeight,
                    "After removal of $toRemove from $list binary tree height increased"
            )
        }
    }

    private fun testRemove2(create: () -> CheckableSortedSet<Int>) {
        val random = Random()
        val binarySet = create()
        val list = mutableSetOf<Int>()

        for (iteration in 1..100) {
            for (i in 1..20) {
                list.add(random.nextInt(100))
                binarySet.add(random.nextInt(100))
            }
        }

        //~~~~~~~~//
        val last = list.max()
        val first = list.min()
        assertEquals(last, binarySet.last())
        assertEquals(first, binarySet.first())

        binarySet.remove(last)
        assertFalse(binarySet.contains(last))
        assertTrue(binarySet.contains(first))

        var i = 0
        val oldSize = binarySet.size
        val myRandom = (first!!..(last!! / 2)).random()
        for (j in 0 until binarySet.last() step myRandom)
            if (binarySet.contains(j)) {
                binarySet.remove(j)
                i++
                assertTrue(list.contains(j))
            }
        assertEquals(oldSize - i, binarySet.size)
    }

    @Test
    @Tag("Normal")
    fun testRemoveSGT() {
        testRemove { createKotlinTree() }
        testRemove2 { createKotlinTree() }
    }

    private fun testIterator(create: () -> CheckableSortedSet<Int>) {
        val random = Random()
        for (iteration in 1..100) {
            val list = mutableListOf<Int>()
            for (i in 1..20) {
                list.add(random.nextInt(100))
            }
            val treeSet = TreeSet<Int>()
            val binarySet = create()
            assertFalse(binarySet.iterator().hasNext(), "Iterator of empty set should not have next element")
            for (element in list) {
                treeSet += element
                binarySet += element
            }
            val treeIt = treeSet.iterator()
            val binaryIt = binarySet.iterator()
            println("Traversing $list")
            while (treeIt.hasNext()) {
                assertEquals(treeIt.next(), binaryIt.next(), "Incorrect iterator state while iterating $treeSet")
            }
            val iterator1 = binarySet.iterator()
            val iterator2 = binarySet.iterator()
            println("Consistency check for hasNext $list")
            // hasNext call should not affect iterator position
            while (iterator1.hasNext()) {
                assertEquals(
                        iterator2.next(), iterator1.next(),
                        "Call of iterator.hasNext() changes its state while iterating $treeSet"
                )
            }
        }
    }

    @Test
    @Tag("Normal")
    fun testIteratorSGT() {
        testIterator { createKotlinTree() }
    }

    private fun testIteratorRemove(create: () -> CheckableSortedSet<Int>) {
        val random = Random()
        for (iteration in 1..100) {
            val list = mutableListOf<Int>()
            for (i in 1..20) {
                list.add(random.nextInt(100))
            }
            val treeSet = TreeSet<Int>()
            val binarySet = create()
            for (element in list) {
                treeSet += element
                binarySet += element
            }
            val toRemove = list[random.nextInt(list.size)]
            treeSet.remove(toRemove)
            println("Removing $toRemove from $list")
            val iterator = binarySet.iterator()
            var counter = binarySet.size
            while (iterator.hasNext()) {
                val element = iterator.next()
                counter--
                print("$element ")
                if (element == toRemove) {
                    iterator.remove()
                }
            }
            assertEquals(
                    0, counter,
                    "Iterator.remove() of $toRemove from $list changed iterator position: " +
                            "we've traversed a total of ${binarySet.size - counter} elements instead of ${binarySet.size}"
            )
            println()
            assertEquals(treeSet.size, binarySet.size, "Size is incorrect after removal of $toRemove from $list")
            for (element in list) {
                val inn = element != toRemove
                assertEquals(
                        inn, element in binarySet,
                        "$element should be ${if (inn) "in" else "not in"} tree"
                )
            }
            assertTrue(binarySet.checkInvariant(), "Binary tree invariant is false after tree.iterator().remove()")
        }
    }

    @Test
    @Tag("Hard")
    fun testIteratorRemoveSGT() {
        testIteratorRemove { createKotlinTree() }
    }

}